requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
python-dotenv
# For testing
pytest
